Calibration code for rover
starts up, during calibration, the magnetometer sensor needs to move around, I suggest it moves to every possible area in its range of motion so that it can handle all the possible magnetic interferences from the environment inside and out. 
It is currently only made to move around the yaw direction using the rover servo motors.