ethernet code if computer 2 is a mac
ethernet code is the code that the second computer needs to run to connect to the router PC to receive the transmissions; this is the version used during the demo. What it does is open its port and listen for data. Whenever it receives it, it prints it to verify data is being sent